platform/surface: aggregatgor_registry: Add HID devices for sensors and USCI client to SP8
Add software nodes for the HID sensor collection and the UCM USCI HID
client to the Surface Pro 8. In contrast to the type-cover devices,
these devices are directly attached to the SAM controller, without any
hub.

This enables support for HID-based sensors, including the ones used for
automatic screen rotation, on the Surface Pro 8.

Signed-off-by: Maximilian Luz <luzmaximilian@gmail.com>