fs: do not set no_llseek in fops
The struct file_operations ->llseek() method is supposed to be called in
two places: vfs_llseek() and dump_skip(). Both treat NULL and no_llseek
as equivalent.

The value of ->llseek is also examined in __full_proxy_fops_init() and
ovl_copy_up_data(). For the former we could as well treat no_llseek as
NULL; no need to do a proxy wrapper around the function that fails with
-ESPIPE without so much as looking at its arguments.  Same for the
latter - there no_llseek would end up with skip_hole set true until the
first time we look at it. At that point we call vfs_llseek(), observe
that it has failed (-ESPIPE), shrug and set skip_hole false. We might as
well have done that from the very beginning.

In other words, any place where .llseek is set to no_llseek could just
as well set it to NULL.

Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>