clocksource: Permit limited-duration clocksource watchdogging
Some systems want regular clocksource checking, but their workloads
cannot tolerate the overhead of full-time clocksource checking.
Therefore, add a CLOCKSOURCE_WATCHDOG_DURATION Kconfig option and a
clocksource.watchdog_duration kernel-boot parameter that limits the
clocksource watchdog to the specified number of minutes past boot.
Values of zero disable checking, and a value of -1 restores the
traditional behavior of always checking.

This does change behavior compared to older kernels, but recent kernels
disable the clocksource watchdog completely in the common case where the
TSC is judged to be trustworthy.  This change in behavior is therefore
not a real issue.

Link: https://lore.kernel.org/all/a82092f5-abc8-584f-b2ba-f06c82ffbe7d@redhat.com/
Reported-by: Waiman Long <longman@redhat.com>
Reported-by: Feng Tang <feng.tang@intel.com>
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Cc: Thomas Gleixner <tglx@linutronix.de>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Borislav Petkov <bp@alien8.de>
Cc: Dave Hansen <dave.hansen@linux.intel.com>
Cc: "H. Peter Anvin" <hpa@zytor.com>
Cc: Daniel Lezcano <daniel.lezcano@linaro.org>
Cc: Feng Tang <feng.tang@intel.com>
Cc: Waiman Long <longman@redhat.com
Cc: <x86@kernel.org>