ACPI: Drop the custom_method debugfs interface
The ACPI custom_method debugfs interface is security-sensitive and
concurrent access to it is broken [1].

Moreover, the recipe for preparing a customized version of a given
control method has changed at one point due to ACPICA changes, which
has not been reflected in its documentation, so whoever used it before
has had to adapt an no problems with it have been reported.

The latter likely means that the number of its users is limited at best
and attempting to fix the issues mentioned above is likely not worth the
effort.  Moreover, if it gets broken in the process, the breakage may not
be readily discovered, so deleting it altogheher appeares to be a better
option.

Accordingly, drop custom_method along with its (outdated anyway)
documentation.

Link: https://lore.kernel.org/linux-acpi/20221227063335.61474-1-zh.nvgt@gmail.com/ # [1]
Reported-by: Hang Zhang <zh.nvgt@gmail.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Reviewed-by: Zhang Rui <rui.zhang@intel.com>