kbuild: deb-pkg: default dpkg-buildpackage --build
deb-pkg's dpkg-buildpackage invocation was added --build=source,binary
recently. Before, it relied on the default, which is the same, so it
no difference for the built packages.

However, the explicit --build eliminates overriding it via DPKG_FLAGS,
which used to fill the gap of generating only a source package without
building it.

Recover the old, default behavior.

Fixes: 7bf4582d7aad ("kbuild: deb-pkg: create source package without cleaning")
Signed-off-by: Bastian Germann <bage@linutronix.de>
Reviewed-by: Nicolas Schier <nicolas@fjasle.eu>
Signed-off-by: Masahiro Yamada <masahiroy@kernel.org>