mm: userfaultfd: avoid passing an invalid range to vma_merge()
The userfaultfd_[un]register() functions will knowingly pass an invalid
address range to vma_merge(), then rely on it failing to merge to indicate
that the VMA should be split into a valid one.

This is not something that should be relied upon, as vma_merge()
implicitly assumes in cases 5-8 that curr->vm_start == addr.  This is now
enforced since commit b0729ae0ae67 ("mm/mmap/vma_merge: explicitly assign
res, vma, extend invariants") with an explicit VM_WARN_ON() check.

Since commit 29417d292bd0 ("mm/mmap/vma_merge: always check invariants")
this check is performed unconditionally, which caused this assert to arise
in tests performed by Mark [1].

This patch fixes the issue by performing the split operations before
attempting to merge VMAs in both instances.  The problematic operation is
splitting the start of the VMA since we were clamping to the end of the
VMA in any case, however it is useful to group both of the split
operations together to avoid egregious goto's and to abstract the code
between the functions.

As well as fixing the repro described in [1] this also continues to pass
uffd unit tests.

[1]:https://lore.kernel.org/all/ZFunF7DmMdK05MoF@FVFF77S0Q05N.cambridge.arm.com

Link: https://lkml.kernel.org/r/20230515193232.67552-1-lstoakes@gmail.com
Fixes: 29417d292bd0 ("mm/mmap/vma_merge: always check invariants")
Signed-off-by: Lorenzo Stoakes <lstoakes@gmail.com>
Reported-by: Mark Rutland <mark.rutland@arm.com>
Closes: https://lore.kernel.org/all/ZFunF7DmMdK05MoF@FVFF77S0Q05N.cambridge.arm.com/
Reviewed-by: Liam R. Howlett <Liam.Howlett@oracle.com>
Cc: Alexander Viro <viro@zeniv.linux.org.uk>
Cc: Christian Brauner <brauner@kernel.org>
Cc: Mike Rapoport (IBM) <rppt@kernel.org>
Cc: Peter Xu <peterx@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>